Fix for CI errors
